pyproject.toml
[tool.poetry]
name = "base-bot"
version = "0.1.0"
description = "Base bot for telegram projects"
authors = ["lurian <lurian-code@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
python-telegram-bot = {extras = ["all"], version = "^20.1"}
pydantic = "^1.10.6"
motor = "^3.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.20.3"
pytest-env = "^0.8.1"
black = "^23.7.0"
debugpy = "^1.6.7"
asyncio = "^3.4.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
